+++ builds/OneeChan.meta.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name         OneeChan
-// @version      5.2.0
+// @version      5.2.1
 // @namespace    OneeChan
 // @description  Customizable rice and themes for 4chan and 4chan X.
 // @license      GPLv3; https://github.com/Nebukazar/OneeChan/blob/master/LICENSE